+++ pkg/types/apiResponse.go
@@ -1,10 +1,10 @@
 package types
 
 type DayLevelFoodConsumption struct {
-	Name       string `json:"name"`
-	Kcal       uint   `json:"kcal"`
-	Protein    uint   `json:"protein"`
-	Fiber      uint   `json:"fiber"`
-	Quantity   uint   `json:"quantity"`
-	Created_at string `json:"CreatedAt"`
+	Name       string  `json:"name"`
+	Kcal       float32 `json:"kcal"`
+	Protein    float32 `json:"protein"`
+	Fiber      float32 `json:"fiber"`
+	Quantity   float32 `json:"quantity"`
+	Created_at string  `json:"CreatedAt"`
 }
